requirements.txt
# Core
openai>=1.50.0
chromadb>=0.5.0
python-dotenv>=1.0.0

# Data
pandas>=2.0.0
requests>=2.31.0
tiktoken>=0.7.0

# LangChain
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Evaluation
ragas>=0.2.0
datasets>=2.20.0
sentence-transformers>=3.0.0
